creating a module for whatsapp shared stuff
