--- corehq/messaging/smsbackends/turn/exceptions.py
@@ -1,2 +1,0 @@
-class WhatsAppTemplateStringException(Exception):
-    pass
